+++ frontend/src/hooks/useFetchEngineers.ts
@@ -6,6 +6,7 @@ type TEngineers = {
     unique_id: string;
     engineer_firstname: string;
     engineer_lastname: string;
+    repairshopr_id: number;
     department: string;
 }[];