+++ frontend/src/lib/vehicle_checklist_table_columns.ts
@@ -13,5 +13,17 @@ const columns: TColumns = [
         header: "Created",
         accessorKey: "formatted_created_at",
     },
+    {
+        header: "Next service",
+        accessorKey: "formatted_next_service_date",
+    },
+    {
+        header: "KMs before",
+        accessorKey: "mileage",
+    },
+    {
+        header: "KMs after",
+        accessorKey: "mileage_after",
+    },
 ];
 export default columns;