+++ server/controllers/driver_app/get_checklist.js
@@ -3,7 +3,7 @@ import { pool } from "../../db.js";
 const getChecklists = async (req, res) => {
     try {
         const { rows } = await pool.query(
-            "SELECT *, TO_CHAR(created_at, 'YYYY-MM-DD') AS formatted_created_at from vehicle_checklist order by created_at desc"
+            "SELECT *, TO_CHAR(created_at, 'YYYY-MM-DD') AS formatted_created_at, COALESCE(DATE(next_service_date)::text, '') as formatted_next_service_date from vehicle_checklist order by created_at desc"
         );
 
         res.json(rows);