+++ server/controllers/engineers/get_engineers.js
@@ -3,7 +3,7 @@ import { pool } from "../../db.js";
 const getEngineers = async (req, res) => {
     try {
         const { rows } = await pool.query(
-            "SELECT id, unique_id, engineer_firstname ,engineer_lastname, department from engineers order by created_at desc"
+            "SELECT id, unique_id, engineer_firstname ,engineer_lastname, repairshopr_id, department from engineers order by created_at desc"
         );
 
         res.json(rows);